Resolving errors in the load configuration

+++ src/main/java/com/IceHockeyLeague/LeagueManager/Team/ITeam.java
@@ -41,5 +41,5 @@ public interface ITeam {
 
     double calculateTeamStrength(ITeamStrength teamStrength);
 
-    void setTeamStrength(int strength);
+    void setTeamStrength(float strength);
 }
